+++ web/src/app/[locale]/(auth)/loading.tsx
@@ -1,5 +1,5 @@
 import Loading from "@/components/Loading";
 
 export default function Fallback() {
-  return <Loading stage="/[locale]/loading.tsx" />;
+  return <Loading stage="/[locale]/(auth)/loading.tsx" />;
 }
