+++ web/src/app/[locale]/(auth)/settings/loading.tsx
@@ -1,0 +1,5 @@
+import Loading from "@/components/Loading.tsx";
+
+export default function Fallback() {
+  return <Loading stage="/[locale]/(auth)/settings/loading.tsx" />;
+}
