+++ web/src/components/Avatar.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Image from "next/image";
 import { useEffect, useRef, useState } from "react";
 export default function Avatar({
