+++ web/src/next/types.ts
@@ -1,0 +1,2 @@
+// next can't type searchParams ew
+export type SearchParams = Promise<{ [key: string]: string | string[] | undefined }>;
